Frontend and backend

+++ routes/cc1_route.js
@@ -34,10 +34,13 @@ var types = ["hr", "hod", "ci", "cc", "am"];
 
 const router = express.Router();
 var key = process.env.PRIVATEKEY;
+const cors = require('cors');
+router.use(cors());
+let tokenString;
 var verifyToken =async function (req, res, next) {
 
     try {
-        var tokenString = jwt.verify(req.headers.token, key);
+         tokenString = jwt.verify(req.headers.token, key);
         req.id = tokenString.id;
         req.userType = tokenString.userType;
         if (tokenString.userType != "cc") {
@@ -60,16 +63,16 @@ router.use(verifyToken);
 
 router.route("/viewslotlinking").get(async(req,res)=>{
     try{
-    var id=req.id;
-    var cours=await courses.find({cordinatorid:id},{_id:0}).select('courseid');
-    //courses which this cord gives
-    var cordcourses=[];
-    for(var i=0;i<cours.length;i++){
-        cordcourses.push(cours[i].courseid);
-    }
+    var id=tokenString.id;
+    // var cours=await courses.find({cordinatorid:id},{_id:0}).select('courseid');
+    // //courses which this cord gives
+    // var cordcourses=[];
+    // for(var i=0;i<cours.length;i++){
+    //     cordcourses.push(cours[i].courseid);
+    // }
     //i saved 2 requests with the course
     
-    var results=await requests.SlotLinkingRequest.find({course:{$in:cordcourses}},{_id:0});
+    var results=await requests.SlotLinkingRequest.find({destinationid:id,status:"pending"},{_id:0});
 
     res.send(results);
     }
@@ -84,29 +87,32 @@ router.route("/viewslotlinking").get(async(req,res)=>{
 router.route("/acceptslotlinking").post(async(req,res)=>{
     try{
         const schema=joi.object({
-            id:joi.string().required(),
-            senderid:joi.string().required(),
-            course:joi.string().required()
+            id:joi.string().required(),//reqid
+             //answer:joi.string().required(),
+            // course:joi.string().required()
         });
         const {error,value}=schema.validate(req.body);
         if(error!=undefined){
             res.send(error.details[0].message);
             return;
         }
-        const r=await requests.SlotLinkingRequest.findOne(value);
+        const r=await requests.SlotLinkingRequest.findOne({id:req.body.id});
         if(r==undefined){
             res.send("cannot find the request");
         }
         const slot=await courseschedule.findOne({courseid:r.course,location:r.location,slot:r.slot,day:r.day});
+        if (slot) {
+            if(slot.userid==undefined){
+                await courseschedule.findOneAndUpdate({courseid:r.course,location:r.location,slot:r.slot,day:r.day},{userid:r.senderid});
+                await requests.SlotLinkingRequest.findOneAndUpdate(value,{status:"accepted"});
+                res.send("request accepted and slot has been updated")
+            }else{
+                res.send("this slot is already taken by another ta or it dosent exist");
 
-        if(slot.userid==undefined){
-            await courseschedule.findOneAndUpdate({courseid:r.course,location:r.location,slot:r.slot,day:r.day},{userid:r.senderid});
-            await requests.SlotLinkingRequest.findOneAndUpdate(value,{status:"accepted"});
-            res.send("request accepted and slot has been updated")
-        }else{
-            res.send("this slot is already taken by another ta ");
-
-        }
+            }
+            
+        }res.send("slot dosent exist");
+        
     }catch(err){
         console.log(err);
         res.send("an error has occurred");
@@ -118,8 +124,8 @@ router.route("/rejectslotlinking").post(async(req,res)=>{
     try{
         const schema=joi.object({
             id:joi.string().required(),
-            senderid:joi.string().required(),
-            course:joi.string().required()
+            // senderid:joi.string().required(),
+            // course:joi.string().required()
         });
         const {error,value}=schema.validate(req.body);
         if(error!=undefined){
@@ -127,7 +133,7 @@ router.route("/rejectslotlinking").post(async(req,res)=>{
             return;
         }
         const r=await requests.SlotLinkingRequest.findOne(value);
-        if(r==undefined){
+        if(!r){
             res.send("cant find this request");
             return;
         }
@@ -158,16 +164,22 @@ router.route("/editcourseschedule")
         day:joi.string().required(),
         slot:joi.number().required(),
         location:joi.string().required(),
-        userid:joi.string()
+        //userid:joi.string()
     });
     const {error,value}=schema.validate(req.body);
     if(error!=undefined){
         res.send(error.details[0].message);
         return;
     }
-    var newslot= new courseschedule(value);;
+    let exist=await courses.findOne({courseid:req.body.courseid})
+    if (exist) {
+        var newslot= new courseschedule(value);;
     await newslot.save();
     res.send("slot added")
+    } else {
+        res.send("no such course")
+    }
+    
 }
 catch(err){
     console.log(err)
@@ -184,15 +196,26 @@ catch(err){
             day:joi.string().required(),
             slot:joi.number().required(),
             location:joi.string().required(),
-            userid:joi.string()
+            nday:joi.string().required(),
+            nslot:joi.number().required(),
+            nlocation:joi.string().required(),
+           // userid:joi.string()
         });
         const {error,value}=schema.validate(req.body);
         if(error!=undefined){
             res.send(error.details[0].message);
             return;
         }
-        await courseschedule.findOneAndUpdate({courseid:value.courseid},value);
-        res.send("slot updated")
+        console.log(req.body)
+        let exist=await courses.findOne({courseid:req.body.courseid})
+
+        
+        if (exist) {
+            await courseschedule.findOneAndUpdate({courseid:req.body.courseid,day:req.body.day,slot:req.body.slot,location:req.body.location},{day:req.body.nday,slot:req.body.nslot,location:req.body.nlocation});
+        res.send("slot UPDATED")
+        } else {
+            res.send("no such data")
+        }
     }
     catch(err){
         console.log(err)
@@ -208,15 +231,21 @@ catch(err){
             day:joi.string().required(),
             slot:joi.number().required(),
             location:joi.string().required(),
-            userid:joi.string()
+           // userid:joi.string()
         });
         const {error,value}=schema.validate(req.body);
         if(error!=undefined){
             res.send(error.details[0].message);
             return;
-        }
-        await courseschedule.deleteOne(value);
+        }let exist=await courses.findOne({courseid:req.body.courseid})
+        let exist1=await courseschedule.findOne(value)
+        if (exist&&exist1) {
+            await courseschedule.deleteOne(value);
         res.send("slot deleted")
+        } else {
+            res.send("no such course or slot")
+        }
+
     }
     catch(err){
         console.log(err)
